refactor(113176): adiciona field array contratos em outro arquivo
